// 06032026/sce-front/src/app/model/autorizacionNacionBean.ts
export class AutorizacionNacionBean{
    id: number;
    numeroAutorizacion: number;
    estadoAprobacion: string;
    descEstadoAprobacion: string;
    autorizacion: number;
    tipoAutorizacion: string;
    descTipoAutorizacion: string;
    detalle:  string;
    codigoCentroComputo: string;
    nombreCentroComputo: string
    tipoDocumento: string;
    descTipoDocumento: string;
    idDocumento: number;
    usuarioCreacion: string;
    fechaCreacion: string;
    usuarioModificacion: string;
    fechaModificacion: string;
}